requirements.txt
unsloth
transformers
datasets
accelerate
peft
trl
bitsandbytes
sentencepiece
pyyaml
rich